+++ src/test/java/domain/ApplicationTest.java
@@ -1,14 +1,12 @@
-package lotto;
-
-import camp.nextstep.edu.missionutils.test.NsTest;
-import domain.Application;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
+package domain;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
+
+import camp.nextstep.edu.missionutils.test.NsTest;
+import java.util.List;
+import org.junit.jupiter.api.Test;
 
 class ApplicationTest extends NsTest {
     private static final String ERROR_MESSAGE = "[ERROR]";
