+++ src/test/java/domain/lotto/LottosTest.java
@@ -1,11 +1,8 @@
-package lotto;
+package domain.lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.winningLotto.BonusNumber;
-import domain.lotto.Lotto;
-import domain.lotto.LottoNumber;
-import domain.lotto.Lottos;
 import domain.winningLotto.WinningLotto;
 import domain.winningLotto.WinningNumbers;
 import java.util.List;
