+++ src/test/java/domain/lotto/PurchaseAmountTest.java
@@ -1,10 +1,9 @@
-package lotto;
+package domain.lotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import constants.ErrorMessage;
-import domain.lotto.PurchaseAmount;
 import org.junit.jupiter.api.Test;
 
 public class PurchaseAmountTest {
