+++ src/test/java/domain/winningLotto/BonusNumberTest.java
@@ -1,12 +1,10 @@
-package lotto;
+package domain.winningLotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import constants.ErrorMessage;
-import domain.winningLotto.BonusNumber;
 import domain.lotto.LottoNumber;
-import domain.winningLotto.WinningNumbers;
 import java.util.List;
 import org.junit.jupiter.api.Test;
