+++ src/test/java/domain/winningLotto/WinningLottoTest.java
@@ -1,11 +1,8 @@
-package lotto;
+package domain.winningLotto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.winningLotto.BonusNumber;
 import domain.lotto.LottoNumber;
-import domain.winningLotto.WinningLotto;
-import domain.winningLotto.WinningNumbers;
 import java.util.List;
 import org.junit.jupiter.api.Test;
